move BRT specific stuff into brt flow
